Try to ensure marker is not in a group on "/:item" url
the "/:item" url didn't take into account that a marker that is part of
the group can neither be used in a "panTo" nor can it be used to
"openPopup" (both happening during showDetails).
To make sure both can indeed happen use "fitBounds" which choose the
appropriate zoomLevel to make sure the marker is not only the center but
also zoomed in enough.
Unfortunately on the choosen maxZoom=19 there where no tiles, hence I
decreased the maxZoom to 17