Center toolbar and set maxwidth to 800px
It now always shows up in two rows, so that all filter options are shown
in the center of the screen at all times. This should make the filter
options more visible, and avoid overlong horizontal lines.

For small screens (<380px) they are shown on the left, but
on small screens you can't miss them anyway. And showing all of them on
the left makes it more consistent between open and closed dropdowns.
Previously "Kontakt-Art" was shown on the right in the second row and
"Produkte" on the left in the third row. Opening "Kontakt-Art" pushed
"Produkte" down even though they weren't visibly adjacent.