Add locate button and show complete Germany by default
instead of zooming in by default, which means the user gets asked for
his location without knowing what it is for (anti-pattern!) the map now
shows Germany by default. With the new locate button the user has the
possibility to jump to his location.
The home button brings the user always back to the default Germany-wide
zoom level and center position.
If the user denies access to his location, the locate button behaves
just like the home button.